Sloppy implementation of time display using semaphores and mutex.
Created variables to keep track of current location in file to display to screen. However, audioTask() runs on second core to reduce audio latency issues. Used semaphore to alter values when needed. Used in audioTask() to update minutes and seconds. Used in loop() twice: First, when new file is selected, minutes and seconds are reset to 0. Second, when displaying current time, values are pulled.